+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix context menus
+
 ## 0.14.1 - 2022 Aug 31
 - Fix app launch error
 - Fix app exit when launch fails